SC-027 added import for databinding, switched size to double [EB,LC]

+++ app/src/main/java/at/tugraz05/slimcat/AddcatActivity.kt
@@ -219,7 +219,7 @@ class AddcatActivity : AppCompatActivity() {
                 0.0
             }
         }
-        binding.cat!!.size = finalSize.toDouble()
+        binding.cat!!.size = finalSize.toInt()
     }
 
     fun getWeightHintStr():String{
